stage_1/requirements.txt
torch
torchaudio
transformers
vector-quantize-pytorch
pyyaml
tqdm
scikit-learn
joblib